backend/requirements.txt
flask
flask-cors
openai
langchain
langchain-community
faiss-cpu
pdfplumber
python-dotenv
requests
